+++ NEWS.md
@@ -1,5 +1,8 @@
 # mnirs 0.4.2
 
+* Create `analyse_kinetics()` to process mNIRS kinetics on nested `nirs_channels` within a list of data.frames.
+    * Currently implemented for `method = "peak_slope"` only, with other methods coming soon.
+    * Includes generic `print()` function.
 * Remove `{tidyr}` dependency from `plot.mnirs()`, and from `{mnirs}` package dependencies.
 * Update `signif_trailing()` to avoid overprinting digits. Used internally for display.
     * `format = c("max_digits", "max_signif")` will print the lesser of either `digits`, or the maximum decimals/sigfigs in the data.